requirements.txt
flask
langchain
openai
python-dotenv
flask-cors
flask-restful